重构实体类ID类型和对应接口，用户token管理优化
重构实体类ID类型，将Long类型改为String或Integer以适应业务需求
调整服务层和控制器层接口参数类型，确保类型一致性
添加UserContextHolder工具类管理用户上下文
优化数据库字段映射和实体类注解
移除冗余代码并更新相关测试

+++ src/main/java/com/menu/menu/dto/BasketDTO.java
@@ -8,7 +8,7 @@ public class BasketDTO {
 
     @Schema(description = "食材ID", required = true)
     @NotNull(message = "食材ID不能为空")
-    private Long ingredientId;
+    private Integer ingredientId;
 
     @Schema(description = "食材数量", required = true)
     @NotNull(message = "食材数量不能为空")
@@ -18,8 +18,8 @@ public class BasketDTO {
     private String unit;
 
     // Getters and setters
-    public Long getIngredientId() { return ingredientId; }
-    public void setIngredientId(Long ingredientId) { this.ingredientId = ingredientId; }
+    public Integer getIngredientId() { return ingredientId; }
+public void setIngredientId(Integer ingredientId) { this.ingredientId = ingredientId; }
     public Integer getQuantity() { return quantity; }
     public void setQuantity(Integer quantity) { this.quantity = quantity; }
     public String getUnit() { return unit; }
